feat(UserDetails, EditUserModal): divided edit user view into its own component, set date is weird still
